feature/AB#24531-SupplierRemodel Seperate App Service from Supplier
